rename imports in service/controller/kvm/v7patch1: github.com/giantswarm/cluster-operator/service/controller/kvm/v7 > github.com/giantswarm/cluster-operator/service/controller/kvm/v7patch1
