+++ service/controller/kvm/v7patch1/resource/chartconfig/create.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/giantswarm/cluster-operator/pkg/v7/chartconfig"
 	"github.com/giantswarm/cluster-operator/pkg/v7patch1/key"
-	kvmkey "github.com/giantswarm/cluster-operator/service/controller/kvm/v7/key"
+	kvmkey "github.com/giantswarm/cluster-operator/service/controller/kvm/v7patch1/key"
 )
 
 func (r *Resource) ApplyCreateChange(ctx context.Context, obj, createChange interface{}) error {
