+++ service/controller/kvm/v7patch1/resource/configmap/desired.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/giantswarm/cluster-operator/pkg/v7/configmap"
 	"github.com/giantswarm/cluster-operator/pkg/v7patch1/key"
-	kvmkey "github.com/giantswarm/cluster-operator/service/controller/kvm/v7/key"
+	kvmkey "github.com/giantswarm/cluster-operator/service/controller/kvm/v7patch1/key"
 )
 
 func (r *Resource) GetDesiredState(ctx context.Context, obj interface{}) (interface{}, error) {
